Add webkit umbrella header, split WKUserScript
